+++ crates/ruma-client-api/CHANGELOG.md
@@ -5,6 +5,7 @@ Breaking changes:
 * Remove `PartialEq` implementations for a number of types
   * If the lack of such an `impl` causes problems, please open a GitHub issue
 * Split `uiaa::UserIdentifier::ThirdParty` into two separate variants
+* Make `message::get_message_events::v3::Request::new`'s `from` parameter optional
 
 Improvements: